refactor: move pages folder to project root
